refactor: 分離したポート層でAI関連コードを再構成
- AIPortとMoveGeneratorPortを`domain/ai`から`application/ports`に移動
- AI関連型を`domain/ai/types`から`domain/models/ai/types`に移動
- `App.tsx`や各サービスコードで新しいインポートパスを調整
- AI機能のポート層分離により、アーキテクチャの明確性と拡張性を改善
