+++ app/src/infrastructure/di/DefaultContainer.ts
@@ -1,12 +1,12 @@
-import { AIService } from '../../application/ai/AIService'
 import { Container } from '../../application/di/Container'
+import type { AIPort } from '../../application/ports/AIPort.ts'
 import type { GamePort } from '../../application/ports/GamePort'
 import type { InputPort } from '../../application/ports/InputPort'
 import type { StoragePort } from '../../application/ports/StoragePort'
 import type { TimerPort } from '../../application/ports/TimerPort'
+import { AIService } from '../../application/services/AIService.ts'
 import { GameApplicationService } from '../../application/services/GameApplicationService'
 import { InputApplicationService } from '../../application/services/InputApplicationService'
-import type { AIPort } from '../../domain/ai/ports'
 import { ChainDetectionService } from '../../domain/services/ChainDetectionService'
 import { CollisionService } from '../../domain/services/CollisionService'
 import { PuyoSpawningService } from '../../domain/services/PuyoSpawningService'
